Decouple from shared-workflows (#2)
This PR makes the GHA workflows here independent of those in
[rapidsai/shared-workflows](https://github.com/rapidsai/shared-workflows/).
It extracts the necessary logic into self-contained workflows so that we
are not tied to the RAPIDS branching/versioning strategy. In addition,
it splits the core logic into a shared workflow that can be used for
both building and publishing, which will enable downstream work to
actually start publishing packages.