reset VERSION to 1.18.1 (#26)
In recent PRs, I set `VERSION` to a pre-release version to test changes
to wheel-publishing logic here:

* #23 
* #24
* #25

That work is complete and the changes appear to be working
(https://github.com/rapidsai/ucx-wheels/pull/23#discussion_r2089128280).

This reverts `VERSION` back to 1.18.1, the value it was at before these
changes.

This should not overwrite any existing packages on
https://pypi.anaconda.org/rapidsai-wheels-nightly/simple/, because we
pass `--skip-existing` to the `anaconda` CLI at upload time:


https://github.com/rapidsai/gha-tools/blob/6b139b6c08fc499933977b707a1d18a9d50f92d7/tools/rapids-wheels-anaconda-github#L64